tratamento de exceções

+++ src/main/java/com/davidsolli/desafio2/services/AtividadeService.java
@@ -3,10 +3,15 @@ package com.davidsolli.desafio2.services;
 import com.davidsolli.desafio2.DTO.AtividadeDTO;
 import com.davidsolli.desafio2.entities.Atividade;
 import com.davidsolli.desafio2.repositories.AtividadeRepository;
+import com.davidsolli.desafio2.services.exceptions.DatabaseException;
+import com.davidsolli.desafio2.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +22,7 @@ public class AtividadeService {
 
     @Transactional(readOnly = true)
     public AtividadeDTO findById(Integer id) {
-        Atividade atividade = repository.findById(id).get();
+        Atividade atividade = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado!"));
         return new AtividadeDTO(atividade);
     }
 
@@ -37,15 +42,26 @@ public class AtividadeService {
 
     @Transactional
     public AtividadeDTO update(Integer id, AtividadeDTO dto) {
-        Atividade entity = repository.getReferenceById(id);
-        saveEntity(entity, dto);
-        entity = repository.save(entity);
-        return new AtividadeDTO(entity);
+        try {
+            Atividade entity = repository.getReferenceById(id);
+            saveEntity(entity, dto);
+            entity = repository.save(entity);
+            return new AtividadeDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Atividade não encontrada!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Integer id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha na integridade referencial!");
+        }
     }
 
     public void saveEntity(Atividade entity, AtividadeDTO dto) {
